+++ client/src/bin/bot/main.rs
@@ -72,6 +72,7 @@ pub fn make_client(
             username,
             password,
             |_| true,
+            Arc::new(|_| {})
         ))
         .ok()
 }