+++ client/src/lib.rs
@@ -66,7 +66,6 @@ use common_net::{
 use common_state::State;
 use common_systems::add_local_systems;
 use comp::BuffKind;
-use crossbeam_channel::Sender;
 use hashbrown::{HashMap, HashSet};
 use image::DynamicImage;
 use network::{ConnectAddr, Network, Participant, Pid, Stream};
@@ -307,17 +306,11 @@ impl Client {
         username: &str,
         password: &str,
         auth_trusted: impl FnMut(&str) -> bool,
-        init_stage_update: Option<Sender<ClientInitStage>>,
+        init_stage_update: Arc<dyn Fn(ClientInitStage) + Send + Sync>,
     ) -> Result<Self, Error> {
-        let update_stage = |stage: ClientInitStage| {
-            if let Some(updater) = &init_stage_update {
-                let _ = updater.send(stage);
-            }
-        };
-
         let network = Network::new(Pid::new(), &runtime);
 
-        update_stage(ClientInitStage::ConnectionEstablish);
+        init_stage_update(ClientInitStage::ConnectionEstablish);
         let mut participant = match addr {
             ConnectionArgs::Tcp {
                 hostname,
@@ -347,7 +340,7 @@ impl Client {
         let in_game_stream = participant.opened().await?;
         let terrain_stream = participant.opened().await?;
 
-        update_stage(ClientInitStage::WatingForServerVersion);
+        init_stage_update(ClientInitStage::WatingForServerVersion);
         register_stream.send(ClientType::Game)?;
         let server_info: ServerInfo = register_stream.recv().await?;
         if server_info.git_hash != *common::util::GIT_HASH {
@@ -366,7 +359,7 @@ impl Client {
 
         ping_stream.send(PingMsg::Ping)?;
 
-        update_stage(ClientInitStage::Authentication);
+        init_stage_update(ClientInitStage::Authentication);
         // Register client
         Self::register(
             username,
@@ -377,7 +370,7 @@ impl Client {
         )
         .await?;
 
-        update_stage(ClientInitStage::LoadingInitData);
+        init_stage_update(ClientInitStage::LoadingInitData);
         // Wait for initial sync
         let mut ping_interval = tokio::time::interval(Duration::from_secs(1));
         let ServerInit::GameSync {
@@ -401,7 +394,7 @@ impl Client {
             }
         };
 
-        update_stage(ClientInitStage::StartingClient);
+        init_stage_update(ClientInitStage::StartingClient);
         // Spawn in a blocking thread (leaving the network thread free).  This is mostly
         // useful for bots.
         let mut task = tokio::task::spawn_blocking(move || {
@@ -3007,7 +3000,7 @@ mod tests {
             username,
             password,
             |suggestion: &str| suggestion == auth_server,
-            None,
+            Arc::new(|_| {}),
         ));
         let localisation = LocalizationHandle::load_expect("en");